some changes
Added new opcodes: `pushspadd` and `pushspsub` which take the current sp and adds or subtracts a number with it and pushes the result to stack. This allows more seamless addressing with the stack. There's two similar opcodes that use the bp frame pointer as well, thus easier to retrieve function arguments from the frame pointer's relative address!
Fixed an overlooked error from `Tagha_read_short` returning a float instead of a short.
updated the primitive assembler for the newest opcodes.