header + assembler changes
Changed the order of the entry point address for scripts from first to last (after natives table).
improved the python and c-based assembler a little with higher level functions such as the natives writer, which takes variadic string literals of the native name and writes them to the binary. pretty helpful to me at least.