Stack Down update
Reoriented the stack so that it grows downward rather than upward. It only makes sense to make it upward when the stack is big endian, but it's supposed to be little endian.
Fixed incorrect value printed from `pushbpadd`.
Updated bytecode for the stack reorientation.
Replaced alot of union-based code with pointer casting for the stack memory.
Finally resolves issue #6 .
Added new bytecode tests: `printf` and `int2char`.
Changed some push and popping functions to return 0 or just return incase of stack overflow or underflow instead of just forcefully exiting the host application, gotta change some others to not exit as well.