Some regression and additions
Removed the API for calling a function inside a script from the host (for now).
Added new opcode called 'jmps' which uses an absolute address as a jump instruction instead of using an immediate value as an address offset.
Temporarily replaced the VMs script vector back to a normal script pointer. Planning to eventually replace the vector with a dictionary so we can access scripts by name.
The purpose for this regression is to take a step back and realize what is needed for the scripts before I abstract the scripting system further, it's better to finalize the script structure and scripting features before finalizing the system.
Added a boolean type for native params and return value pointer.
Replaced some "printf" code with "puts".