Argc and Argv update!
Executing a script now allows you to pass in arguments to script's main function! to keep it compatible with 32-bit and 64-bit archs, the argv array is a union CValue array instead of an array of strings.
Optimized stack pushing code.
Created test bytecode to retrieve and use "argc" and "argv".
Updated the TBC bytecode executables for the latest changes.
Updated the C++ wrapper code for the latest changes.